HACS support
Tentative d'ajout du support HACS
