+++ README.md
@@ -1,10 +1,6 @@
 # Hilo
 Hilo integration for Home Assistant
 
-#Changelog: 0.2
-- Added support for switches
-- Added binary sensor for Hilo Event (Défi Hilo)
-
 # What This Is:
 This is a custom component to allow control of Hilo devices from Home Assistant.
